refactor: extract cloud-dispatch.ts from run.ts (#447)
Step 4: Extract self-contained cloud worker dispatch and polling logic.
